# lib/api-zod/src/generated/types/cancelRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CancelRequest {
  /** @minLength 1 */
  clientId: string;
  /** @minLength 1 */
  tag: string;
}