# lib/api-zod/src/generated/types/scheduleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleRequest {
  /** @minLength 1 */
  clientId: string;
  /** @minLength 1 */
  tag: string;
  /** @minLength 1 */
  title: string;
  body: string;
  url?: string;
  fireAt: Date;
}