# lib/api-zod/src/generated/types/subscribeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WebPushSubscription } from "./webPushSubscription";

export interface SubscribeRequest {
  /** @minLength 1 */
  clientId: string;
  subscription: WebPushSubscription;
}