# lib/api-zod/src/generated/types/webPushSubscription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PushSubscriptionKeys } from "./pushSubscriptionKeys";

export interface WebPushSubscription {
  /** @minLength 1 */
  endpoint: string;
  keys: PushSubscriptionKeys;
}